split read parsing into separate module
